Make debug_log! work on the pico by using defmt::println!
This works like the other implementations, so it will allocate on the heap, format with core::fmt
, have no compile-time switch and not print the location.

It would be nicer to forward to defmt directly, but that means the regular
formatting wouldn't work and defmt needs to be a dependency at the call site.

For more fine-grained debugging it's easier to introduce local defmt usage.